Fix code folding when folding ends on the last line
Previously, a final FoldBlock that begins one line after the last
block and ends at the last line was always created. If the last block
ended on the last line of the listing, this resulted in an
ArrayIndexOutOfBoundException.

This commit updates the converter to only create the final FoldBlock if
it's starting line is on or before the last line of the listing.

Fixes gh-26